custom-elements/instantiating-custom-elements/extensions-to-document-interface/created-callback-create-element.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback should be called if createElement is called</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Created callback should be called if createElement is called">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#extensions-to-document-interface-to-instantiate">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var proto = newHTMLElementPrototype();
    var name = 'x-a';

    doc.registerElement(name, {prototype: proto});
    var customElement = doc.createElement(name);
    assert_equals(proto.createdCallbackCalledCounter, 1,
        'Callback created should be called if createElement is called');
}, 'Test that created callback is called if createElement is called. Type extension is not specified in ' +
    'createElement call');


test(function() {
    var doc = newHTMLDocument();
    var proto = newHTMLElementPrototype();
    var name = 'x-b';

    doc.registerElement(name, {prototype: proto});
    var customElement = doc.createElement(name, name);
    assert_equals(proto.createdCallbackCalledCounter, 1,
        'Callback created should be called if createElement is called');
}, 'Test that created callback is called if createElement is called. Type extension is specified in ' +
  'createElement call');


test(function() {
    var doc = newHTMLDocument();
    var proto = newHTMLElementPrototype();
    var name = 'x-c';
    var customElement = doc.createElement(name);
    assert_equals(proto.createdCallbackCalledCounter, 0,
        'Callback created should not be called if custom element is unresolved');

    doc.registerElement(name, {prototype: proto});
    assert_equals(proto.createdCallbackCalledCounter, 1,
        'Callback created should be called after custom element is registered');
}, 'Test that created callback is called if createElement is called. ' +
    'Custom element is unresolved at first');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(tagName) {
        var obj = doc.createElement(tagName);
        var name = 'x-d-' + tagName;
        var proto = newHTMLElementPrototype();
        doc.registerElement(name, {prototype: proto, extends: tagName});
        var customElement = doc.createElement(tagName, name);

        assert_equals(proto.createdCallbackCalledCounter, 1,
            'Callback created should be called if createElement is called');
    });
}, 'Test that created callback is called if createElement is called. Test extensions of all HTML5 elements');
</script>
</body>
</html>
